customer_support/requirements.txt
sentence-transformers
numpy
faiss-cpu
openai
mlflow
pydantic
python-dotenv
pyyaml
pytest